Move rpmVersionCompare() to headerutil.c for better layer match
No functional change, but we'll need this in the next step.
