+++ lib/rpmvercmp.c
@@ -136,23 +136,3 @@ int rpmvercmp(const char * a, const char * b)
     if (!*one) return -1; else return 1;
 }
 
-int rpmVersionCompare(Header first, Header second)
-{
-    /* Missing epoch becomes zero here, which is what we want */
-    uint32_t epochOne = headerGetNumber(first, RPMTAG_EPOCH);
-    uint32_t epochTwo = headerGetNumber(second, RPMTAG_EPOCH);
-    int rc;
-
-    if (epochOne < epochTwo)
-	return -1;
-    else if (epochOne > epochTwo)
-	return 1;
-
-    rc = rpmvercmp(headerGetString(first, RPMTAG_VERSION),
-		   headerGetString(second, RPMTAG_VERSION));
-    if (rc)
-	return rc;
-
-    return rpmvercmp(headerGetString(first, RPMTAG_RELEASE),
-		     headerGetString(second, RPMTAG_RELEASE));
-}
